Replaces gulp-based JS transpilation, compilation and minification with 11ty’s terser-based inline minified JS tip. Removes babel as a top-level dependency. Removes polyfill.io. Removes IIFEs. Introduces script type=module so that the JS is only available when native browser support for ES6 modules is available.
