--- _includes/js/fuzzylogic.js
@@ -1,1 +1,0 @@
-"use strict";function _createForOfIteratorHelper(e,t){var r;if("undefined"==typeof Symbol||null==e[Symbol.iterator]){if(Array.isArray(e)||(r=_unsupportedIterableToArray(e))||t&&e&&"number"==typeof e.length){r&&(e=r);var n=0,t=function(){};return{s:t,n:function(){return n>=e.length?{done:!0}:{done:!1,value:e[n++]}},e:function(e){throw e},f:t}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var o,a=!0,i=!1;return{s:function(){r=e[Symbol.iterator]()},n:function(){var e=r.next();return a=e.done,e},e:function(e){i=!0,o=e},f:function(){try{a||null==r.return||r.return()}finally{if(i)throw o}}}}function _unsupportedIterableToArray(e,t){if(e){if("string"==typeof e)return _arrayLikeToArray(e,t);var r=Object.prototype.toString.call(e).slice(8,-1);return"Map"===(r="Object"===r&&e.constructor?e.constructor.name:r)||"Set"===r?Array.from(e):"Arguments"===r||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(r)?_arrayLikeToArray(e,t):void 0}}function _arrayLikeToArray(e,t){(null==t||t>e.length)&&(t=e.length);for(var r=0,n=new Array(t);r<t;r++)n[r]=e[r];return n}!function(){for(var e=document.querySelectorAll(".grid"),t=0;t<e.length;t++)!function(n){var e,t,o;"ResizeObserver"in window&&!CSS.supports("width","min(20rem, 100%)")&&(e=n.dataset.min,(t=document.createElement("div")).style.width=e,n.appendChild(t),o=t.offsetWidth,n.removeChild(t),new ResizeObserver(function(e){var t=_createForOfIteratorHelper(e);try{for(t.s();!(r=t.n()).done;){var r=r.value.contentRect.width>o;n.classList.toggle("aboveMin",r)}}catch(e){t.e(e)}finally{t.f()}}).observe(n))}(e[t])}(),function(){function a(){for(;u.firstChild;)u.removeChild(u.firstChild)}var i=null,t=document.querySelector(".search-ui"),u=document.querySelector(".search-results"),r=document.querySelector("#search-str");document.addEventListener("click",function(e){event.target.closest("#nav-search")&&(e.preventDefault(),fetch("/search.json").then(function(e){return e.json()}).then(function(e){i=e.search}),t.classList.toggle("invisible"),r.focus(),r.addEventListener("keyup",function(e){var t=r.value;2<t.length?function(e){e=e.toLowerCase();var t,r=[];for(t in i)-1!=i[t].text.indexOf(e)&&r.push(i[t]);for(t in a(),r){var n=document.createElement("li"),o=document.createElement("a");o.textContent=r[t].title,o.setAttribute("href",r[t].url),n.appendChild(o),u.appendChild(n)}}(t):a()}))},!1)}();
