+++ gulpfile.js
@@ -1,19 +1,6 @@
 const gulp   = require("gulp");
 const sass   = require("gulp-sass");
-const concat = require('gulp-concat');
-const babel  = require('gulp-babel');
-const uglify = require('gulp-uglify');
 const del = require('del');
-
-
-gulp.task('default', () =>
-  gulp.src('src/app.js')
-      .pipe(babel({
-          presets: ['@babel/env']
-      }))
-      .pipe(gulp.dest('dist'))
-);
-
 
 /*
   clear out eleventy-generated files so that no old blog posts are hanging around and showing up where they shouldn’t
@@ -39,29 +26,10 @@ gulp.task('css', function() {
 
 
 /*
- Uglify our javascript files into one.
- Use pump to expose errors more usefully.
-*/
-
-gulp.task('js', function() {
-  return gulp.src("./js/**/*.js")
-    .pipe(concat('fuzzylogic.js'))
-    .pipe(babel({
-      presets: ['@babel/env']
-    }))
-    .pipe(uglify())
-    .pipe(gulp.dest('./_includes/js'));
-});
-
-
-
-
-/*
   Watch folders for changes
 */
 gulp.task("watch", function() {
-  gulp.watch('./scss/**/*.scss', gulp.parallel('css'));
-  gulp.watch('./js/**/*.js', gulp.parallel('js'));
+  gulp.watch('./scss/**/*.scss', gulp.parallel('css'))
 });
 
 
@@ -69,6 +37,5 @@ gulp.task("watch", function() {
   Let's build this sucker.
 */
 gulp.task('build', gulp.parallel(
-  'css',
-  'js'
+  'css'
 ));
